refactor: UserTest 및 관련 코드 Users로 변경
1. UserTest 클래스를 UsersTest로 이름 변경한다.
2. 관련 패키지 및 import 경로를 users로 수정한다.
3. UserCreateRequest 사용 부분을 Users.create로 수정한다.

Resolves #11
